pyproject.toml
[project]
name = "hspn"
version = "0.1.0"
description = "deep operator networks for ARL CFD"
readme = "README.md"
requires-python = ">=3.10,<3.11"
authors = [
  { name = "Marco Christiani", email = "mchristiani@sei.cmu.edu" },
  { name="Jasmine Ratchford", email="jratchford@sei.cmu.edu" },
]

dependencies = [
    "h5py>=3.13.0",
    "hydra-core>=1.3.2",
    "rich>=14.0.0",
    "torch==2.7.0",
    "aim>=3.28.0",
]

[project.optional-dependencies]
dev = [
    "netcdf4>=1.7.2",
    "pytest>=8.4.1",
    "pytest-cov>=4.0.0",
    "rq-dashboard>=0.8.2.2",
    "ruff>=0.9.7",
]
hpo = [
    "cloudpickle>=3.1.1",
    "hydra-optuna-sweeper>=1.2.0",
    "hydra-ray-launcher>=0.1.2",
    "hydra-rq-launcher>=1.2.0",
    "rq>=2.2.0",
]

[project.scripts]
hspn-preprocess = "hspn.prepare:main"
hspn-train = "hspn.train:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/*"
]

[tool.hatch.build.targets.wheel]
packages = ["src/hspn"]

[tool.ruff]
preview = true
line-length = 120
ignore = ["N813"]
lint.dummy-variable-rgx = "^_$"
